UPDATED: Added optional step for cleaning the project.
- Rens whole project:
    - Removed node_modules and package-lock.json
    - Updated npm dependencies
    - Installed dependencies again
    - Cleaned npm cache
    - Built projectUPDATED: Changed the step number for initializing a new local git repository.

- Updated step 11 to initialize a new local git repository.

ADDED: Added a step for creating a new remote repository on GitHub.

- Added step 12 for creating a new remote repository on GitHub.

UPDATED: Changed the step number for adding a remote repository.

- to remote repository.

UPDATED: Changed the step number for pushing to the remote repository.

- Updated step 14 to push to repository.

UPDATED: Changed the step number for starting the development server.

- Updated step 15 to start the development server.

UPDATED: Changed the step number for opening localhost.

- Updated step 16 to open localhost.

REMOVED: Removed the duplicated step for cleaning the project at.